backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
httpx==0.28.0
asyncpg==0.30.0
PyJWT[cryptography]==2.10.1
python-dotenv==1.0.0
